// CerealFrontEnd/src/types.ts
export interface Cereal {
  id: number;
  name: string;
  mfr: string;
  type: string;
  calories: number;
  protein: number;
  fat: number;
  sodium: number;
  fiber: number;
  carbo: number;
  sugars: number;
  potass: number;
  vitamins: number;
  shelf: number;
  weight: number;
  cups: number;
  rating: number;
}